Add stack data structure, to hold the maps. Add stacks, and add 1 Stack per map initially. Within svgContainer, wrap another g (.stack, stackS) around g.map Factor g.map translation into mapTransform(). Add : Stacked(), Stacked.prototype.{mapName,portion,mapName_match()}, Stack(),  Stack_add(), Stack.prototype.{.empty,insert,findIndex,remove,delete,move,dropIn,mapStack,mapStackIndex,mapTransform}(), most of these have been stepped through and work ok.
Change stackDropTarget to be the result of .append(g) not .append(rect) (that was the original intention, both worked but want access to the g's classList in dragged).

Add currentDropTarget, storeDropTarget() and access them in dropTargetMouse{Over,Out}(), to know the DropTarget which the cursor is in.
dragged(): if cursor is in top or bottom dropTarget-s, stack the map.
Add jshint curly : false.

Add action/function updatedStacks, to display current Stack info (not working).
To construct text for that, add : {Stacked,Stack}.prototype.{toString,log}(), stacks.log = Stack.log().
Narrow the selector for mouse{over,out} / handleMouse{Over,Out} so it applies only to foreground path-s and not axis path
(by prefixing .foreground > g >  to path).
 cc67954 41082 Mar  3 15:51 frontend/app/components/draw-map.js