Add Map now adds the chr (axis) consistently, getting just 1 chr data. Move the collation of the chr data using findRecord(chromosome ) etc, from controllers/mapview.js to draw-map.js; i.e. the values pushed via dataReceived are mapsToView instead of the chr data; dataObserver() : move dataReceived.popObject() to an Ember.run.later() and use .get(content); factor chrData() from controllers/mapview.js to utils/utility-chromosome, for use also in draw-map.js
Attributes of oa which are used in closures and have a local copy in draw, e.g. x = oa.x, have been
changed to get a fresh copy in the closure, because oa is constant between renders but the variables may not be.

Because this work is not complete and this commit will be pushed and the branch is being followed,
the render-promises feature is disabled by toggling 2 comments; to re-enable :
  controllers/mapview.js : .observes(mapsToView)
  routes/mapview.js: // refreshModel: true

 3bc5480 166337 Aug  3 11:05 frontend/app/components/draw-map.js
 18d5dc0   3242 Aug  3 11:13 frontend/app/controllers/mapview.js
 d77f683   3035 Aug  3 11:13 frontend/app/routes/mapview.js