fix : show axis ticks for child blocks
this refers to ticks at the locations of features which are connected by a path to an adjacent axis and the other end of the path is outside the current zoom interfaval of the other axis.
Add blockO() - uses Stacked.getAxis() to map from child block id to parent axis;  used this in featureLineS(), and also in featureLine2() and featureLine() which are not in use.
 0f39140 226166 Jun 28 18:48 frontend/app/components/draw-map.js