Use updated dragLimit after window resize
This commit fixes this test case : adjust the window to about half the screen width, then refresh the browser tab, use a URL which defines mapsToView or add axes, then expand window to full-screen, now dragging an axis jumps suddenly to a position just left of the second left-most axis.  (no stacking required).   affects all of the axes to the right of that position.
 ca10768  226676 Jun 29 18:05 frontend/app/components/draw-map.js