Use current adjacency set for requesting aliases.
This error was causing an intermittent issue whereby the getPaths is not done for a newly exposed adjacency, and aliased paths are then not drawn.
collateAdjacentAxes() : change to use stack .dataBlocks() instead of .childBlocks(); cache values in dataBlocks[].
Stacked.prototype.dataBlocks(): filter out the reference blocks; data blocks are recognised by having a .namespace;  use this in a new Stack.prototype.dataBlocks() (rename the old unused one to dataBlocks0).
frontend/app/
 b2040e1 193738 Jul 27 13:37  components/draw-map.js
 42386af  37987 Jul 27 16:22  utils/draw/collate-paths.js
 1b36197   3737 Jul 27 16:15  utils/stacks-adj.js
 02c5f6f  56159 Jul 27 16:16  utils/stacks.js