Change data of child block tspans from name to block.
This is in preparation for having a distinct axis title menu, in particular the delete action, for child data blocks, currently it calls axis delete.
From axisTitleChildren() the .titleText() call is moved to .text() function and the rest is dropped.  Change level of titleText() from Stacked to Block; drop the .map() and move .children() call to tspan .data().
frontend/app/
 009ca57 193692 Jul 27 20:32  components/draw-map.js
 79e04c0  56070 Jul 27 20:25  utils/stacks.js