Ensure that during unview-all, child data blocks are unviewed before their parents.
filter blocksUnviewed into children and parents - then unview them in that order.
Accumulate child data blocks whose parent is being unviewed; these will be unviewed before the parents.
In collate-paths, get store from draw-map component not injection; this was causing error : store.peekRecord is not a function.
Fix a reference to d3Features - it is in flowsService not oa.
frontend/app/
 0f7415e 200304 Aug  3 13:56  components/draw-map.js
 c965750  38234 Aug  3 14:43  utils/draw/collate-paths.js
 bf1c278  56614 Aug  3 13:56  utils/stacks.js