Disable the U_alias flow.
Currently the unique alias implementation checks for uniqueness of a path from one block to the other (i.e. it is injective), but does not check the reverse direction (i.e. whether it is bijective).  So this feature is not presented to the user until the reverse check is added.
Add flowsEnableUAlias = false. filterPaths() : if ! flows.U_alias.enabled then put all aliases in flows.alias.  flow-controls.hbs : only display flow if flow.enabled.
frontend/app/
 0461ba1  7449 Aug  6 14:43  services/data/flows-collate.js
 f9f3d88   699 Aug  6 14:42  templates/components/draw/flow-controls.hbs
 19522da 38297 Aug  6 14:43  utils/draw/collate-paths.js