Fix adoption of >1 child axis in one pass
This error presented as : in the title one of the data blocks had black (reference) instead of colour; a duplicate axis was at the same position, causing blurry text.  Getting all blocks at the same time using URL refresh was the method of reproducing this error.
Set .parent = sd; in aBlock instead of a.
frontend/app/
 fc0a76b 205707 Sep 18 18:34  components/draw-map.js
 dcf355d  58880 Sep 18 18:22  utils/stacks.js