Fixes #114.
loadedViewedChildBlocks() : use .get(isLoaded) instead of namespace to check if block is loaded.
The calculation of generationBlocksUnviewed also recognises a child block with block.get(namespace) but it also checks features, which are not optional for child block, so this addition of b.parent to the check is not essential.
frontend/app/
 3019242 206374 Nov 13 15:20  components/draw-map.js
 cb66437   7685 Nov 13 14:47  services/data/block.js