split Feature Search to a separate tab, placed first then Explorer, View, Upload
Split out uses of goto-feature and goto-feature-list from manage-view to form manage-search.
Use a closure action for loadBlock passed to manage-search, and partially changed in left-panel but
the existing non-closure action is stil passed to manage-explorer and manage-view.
Reduce height of Feature Search input textarea from 20 to 10 lines.
frontend/app/
 890d747   830 Dec  6 10:16  components/panel/manage-view.js
 6cc4a08 20707 Dec  6 11:38  styles/app.css
 2f2e8a4  2782 Dec  6 11:18  templates/components/panel/left-panel.hbs
 6f1a6a8   593 Dec  6 11:35  templates/components/panel/manage-search.hbs
 582c3e8  3436 Dec  6 10:16  templates/components/panel/manage-view.hbs