import d3 via npm instead of bower.
On a branch because this is not needed yet.
This was done to access d3-array.group, but that is not in the currently used version anyway, so this is only smoke-tested so far - it starts and displays mapview OK.
This is just an interim step - it recreates the d3 global containing all functions;  the planned change is to import in each js which uses d3, just the required d3 packages.
frontend/
 28f63cf 208742 Jan  5 19:45  app/components/draw-map.js
 cf7d35c   4420 Jan  4 19:29  app/utils/draw/axis.js
 23d267d   7174 Jan  5 19:56  app/utils/draw/flow-controls.js
 572c529    711 Jan  4 17:59  bower.json
 45770ed   1810 Jan  4 18:09  ember-cli-build.js
 f973a8c 820082 Jan  4 18:02  package-lock.json
 0d9e7c5   2218 Jan  4 18:02  package.json
 1e4f980   1343 Jan  4 18:26  vendor/js/divgrid/divgrid.js