data explorer : change computed functions to handle datasets being a promise
datasets was changed to a promise in aa92328; this propagates that change through the CFs which depend on it.  Split data() into dataFG() and datasetFilter().
entry-scope : change args to be consistent with other components referenced by levelComponent : scopeName -> name, blocks -> values.
Change to changeFilterOrGroup() enables it to be used for devel breakpoint;  it is not required now that attribute filterOrGroup is bound to the input radio-button (s).

undo after commit : trace_dataTree back to 1, indent on dataP, RSVP.Promise,
frontend/app/
 9c322e0  1265 Jan 12 17:14  components/panel/filter-group.js
 262f517 14539 Jan 13 20:11  components/panel/manage-explorer.js
 3fc879a  1524 Jan 13 19:56  components/record/entry-values.js
 1f84f50   494 Jan 13 18:45  templates/components/record/entry-blocks.hbs
 60e12ad   595 Jan 13 20:23  templates/components/record/entry-parent.hbs
 02d0ab5   431 Jan 13 20:23  templates/components/record/entry-scope.hbs
 f66c629  1071 Jan 13 19:25  templates/components/record/entry-values.hbs