Undo previous commit 6b7b67b.
the documented method to disable loopback-component-explorer does not seem to work, further testing required; it now looks like the way to disable it in the hosted servers is to remove that default config file component-config.json, leaving the custom config file e.g. component-config.{local,production}.json, to have effect.
 8a80554  3 Jan 17 12:06 backend/server/component-config.json