data explorer : add block-level filtering : scope, namespace.
added namespace (fieldNamespace).
Add filterMatched : filterGroup may apply at both the dataset level and the blocks level, but it makes sense to apply it at only 1 level, so apply it at block level iff it doesn't apply at dataset level.
Use dataTypedTreeFG in place of dataTyped
filtering : move the children of the matched (true) group to the top level of the result, combine the unmatched alongside them.
In grouping, name the undefined group as unmatched.
Add isFilter.
Move the parentAndScope() filtering of annotation from dataTyped() to dataTypedTreeFG(). Change dataTypedTreeFG() result to a PromiseObject because it is used in template. Factor to form addParentAndScopeLevels().
Rename dataFG to dataTypedFG, and re-add the previous dataFG - still used for the un-typed tabs, which are currently retained for comparison (entry-datasets : Genitic Maps : dataWithoutParent, References : parents).
deferred for subsequent commits : block filtering (annotation) done for parent/scope, not yet done for entry-datasets;  block-level grouping.

 b8d32ad   4669 Jan 20 18:54 doc/notes/design_explorer_CF_diagram.html
frontend/app/
 4f91a5a   1307 Jan 20 20:04  components/panel/filter-group.js
 9b97b63    578 Jan 20 20:18  components/panel/filter-groups.js
 f1c2898  23481 Jan 20 20:04  components/panel/manage-explorer.js
 992f999   3642 Jan 19 20:41  components/record/entry-values.js
 2560f9b   2470 Jan 20 20:18  templates/components/panel/filter-group.hbs
 e03c435   4893 Jan 20 18:46  templates/components/panel/manage-explorer.hbs