Rename mongo_functions_alias.js to paths-aggr.js
mongo_functions_alias.js was used as the basis for the first version of paths-aggr.js; this commit drops that version, and renames mongo_functions_alias.js to paths-aggr.js, which maintains more of the history, without the additional file which is basically the same.
Following commit will contain the (minor) changes which were in paths-aggr.js.
(originally committed in 0b0041b)
 4825a1e 5778 Feb 21 09:27 backend/common/utilities/paths-aggr.js (renamed from doc/mongo_functions_alias.js)