axis-tracks : generate dataBlocks from store Blocks instead of stack; handle -ve interval
axis-2d.js : rename dataBlocks() to dataBlocksS() and add a new version which is based on blocks in Ember store instead of stacks Blocks.
axis-tracks.js : in line with change to axis-2d:dataBlocks(), derive trackBlocks() from trackBlocksR() instead of vice versa.
tracksTree() : handle -ve interval (swap) - this was causing createIntervalTree() to overflow stack.

models/block.js : add axis().
data/block.js : axesBlocks() : use ensureAxis();  add dataBlocks().

draw-map.js : factor ensureAxis() and add to axisApi.
The driver for this commit was to update when blocks are added in axis-tracks; that is incomplete, but the interval swap is needed as a hot-fix commit.
frontend/app/ :
 d0f7c87  10543 Sep  6 13:29  components/axis-2d.js
 0c57e6c  24562 Sep  6 16:16  components/axis-tracks.js
 2312316 223195 Sep  6 16:34  components/draw-map.js
 6ce4cad   6126 Sep  6 09:36  models/block.js
 b1d8da7  14140 Sep  6 15:57  services/data/block.js
 731b006  28403 Sep  6 16:39  styles/app.scss
 60ba799   1847 Sep  6 13:29  templates/components/axis-2d.hbs