Return interval length in featuresCounts result when interval is given / using non-auto bucket boundaries.
block-features.js : split boundaries() into binEvenLengthRound() and binBoundaries(), so that lengthRounded can be returned in the result as idWidth (because idWidth is constant in this result, it could be factored out using groupBy - this would require the datum2Location function to access some context - the API result).
data-types.js : for $bucket result which now includes idWidth, instead of adding a custom rectHeight(), modify datum2Location() to calculate the interval from the given lower bound and interval length : idWidth.
 9d19a4d 8023 Feb  9 23:55 backend/common/utilities/block-features.js
 5e1addb 7945 Feb  9 23:00 frontend/app/utils/data-types.js