features table : lookup block from datasetId and scope extracted from selectedFeatures.Chromosome
manage-features.js : add blockIdFromName() to lookup block by datasetId and scope extracted from selectedFeatures.Chromosome, based on blocksByReferenceAndScope;  use this for the block and intersect filters.
This replaces the use (added in 65b9c8e) of feature.Block which is undefined - selectedFeatures contains .Chromosome not .Block, and it is name + scope rather than block Id.   selectedFeatures will change to a Set of features and this can then be replaced with a direct lookup of feature.block.
 952172c  2485 Feb 24 14:36 frontend/app/components/panel/manage-features.js