Change use of .parent in dataset explorer to handle recent changes for cross-server references.
This restores some of the parent hierarchy in the reference/Genome tabs; still some appearing under parent null so further change required.
manage-explorer.js : Use parentName in place of .parent.name (because of the change to .parent from relationship to CP in 40eebb9, it may be undefined if server with reference is not connected) in child1(), dataParentTyped(), parentAndScope().
dataset.js : implement .children(), previously provided via ember-data .parent relationship
frontend/app/ :
 02e21a2  47240 Jun 29 15:52  components/panel/manage-explorer.js	(trace level change not staged)
 43cb798   3553 Jun 29 16:44  models/dataset.js