push the brushed features in the store of the feature's dataBlock, not the store of the axis referenceBlock.
This affects zooming after brushing, with cross-reference data blocks in split axes.
paths-progressive.js : requestBlockFeaturesInterval() : pass the store of the dataBlockIds to pushFeatureField(), not the store of the param blockA which is the referenceBlock of the axis, and may be on a different server to the data blocks of the axis.
 43cfc76  26505 Jul 15 22:37 frontend/app/services/data/paths-progressive.js