add dependency to address block added part of #179
This seems to have made progressive loading slower - possibly due to repeated requsts with equal zoomedDomain, which will be added to #179 and addressed in a separate commit.
draw/block-adj.js : pathsAliasesResultLength() : for #179, added dependency blockAdj.zoomCounter (which is also in dependency paths, but adding it here made a difference); added detail : paths.{direct,alias}.[] (may not be essential).
models/block-adj.js : call_taskGetPaths() : disable the isIdle check - should not be required.
if caught error is TaskCancelation, use lastResult instead of rethrowing the error.
taskGetPaths() : change to .keepLatest().
 9e37e4d 20360 Aug  7 16:47 frontend/app/components/draw/block-adj.js
 38c6bc0 19344 Aug  7 19:32 frontend/app/models/block-adj.js