add transition on path removal, to verify paths re-filter
This animation is mostly for development purposes;  possibly a quieter version would have a role in production; if so, also include d=pathU in the transition so that paths removed at the edge when zooming in move over the edge.
block-adj.js : draw() : add gPathDashAndRemove(), used in draw(), and pathDashTween().
 39e7e16 23654 Aug 13 20:37 frontend/app/components/draw/block-adj.js