re-connect the split-axis width resizer
axis-2d.js : resized() : if resizer is in <foreignObject> then resize the <foreignObject>
setWidth() : select the <rect> in clipPath if not dualAxis (vertical <path> is used instead of <rect> for border).
 send axisWidthResize{,Ended} to attr drawMap instead of parentView (originally parentView was drawMap).
app.scss : if .axis2dResizer is enabled, don't hide .foreignObject.
domElements.js : eltWidthResizable(): if resizer is in foreignObject, use that for d3.mouse call to get position relative to resizable (parent); add relativeParent.
 8ff83a7 13752 Aug 25 21:02 frontend/app/components/axis-2d.js
 e3a0a2a 33474 Aug 24 16:17 frontend/app/styles/app.scss
 551cc7e 11902 Aug 25 18:30 frontend/app/utils/domElements.js