axis zoom/reset : re-connect .on on <rect>
There is still an issue with state update from brush : buttons are not coloured until e.g. mousewheel-zoom in.
Because this registers for clicks on the <rect>, clicking on the <text> within the <rect> has no effect, so next commit will register for <text> clicks also.
brushHelper() : define zoomSwitchRect, and use .nodes()[0] and [1] to select the rect to receive clicks on.
 fa65260 241122 Sep 23 12:37 frontend/app/components/draw-map.js