axis zoom/reset buttons : update x position of buttons when add track
add extra calls to axis1d.showZoomResetButtonXPosn() in :
 axis-2d.js :  axisShowExtend() and setWidth().
 draw-map.js :  updateAxisTitleSize()
It should not be necessary to have all these calls (it is already called in axis-1d extendedEffect() and extendedWidthEffect(), and axis-2d : axisWidthResizeEnded()); once the necessary dependencies are added that should be sufficent and the remainder can be culled.
The call in axisShowExtend() is currently effective in updating when track is added.
frontend/app/components/ :
 abdf6a0  21864 Sep 30 14:32  axis-2d.js
 8fc1af4 241340 Sep 30 12:23  draw-map.js