Convert axis title menu from $.popover() to hbs
Changing from $.popover() to axis-menu.hbs : easier to extend and maintain; also  $.popover is not defined after version upgrade, and it's not worth battling with packaging to access it.
draw-map.js :
comment out .popover(), replacing with .on(click) -> added showMenu(), which sets .menuAxis to axisName
name the menu button action functions and pass them via axisApi.menuActions = {axisDelete, axisFlip, axisPerpendicular, axisExtend}
configureAxisSubTitleMenu() :  if  .popover is undefined, don't call it. This will be changed to hbs also.

axes-1d.js : add alias drawMap.menuAxis.
app.scss : axis title : alter opacity when hover to indicate that it accesses axis-menu.
.axis-menu{,-inner} and .menu-icons : configure layout and colours.
app-header.hbs : draw/axis-menu is not required here - it is instantiated in axes-1d.
axis-menu.hbs : wrap with EmberTooltip (will change to Popover after commit to have click-close).
Rename .popover-content to .menu-icons and move it to the bottom, inline; drop the div.popover wrap - not needed, it was part of the DOM of the previous axis title menu, copied as a seed for the hbs.

frontend/app/ :
 4dcb2b5 242139 Nov 24 13:36  components/draw-map.js
 6369359   1069 Nov 24 13:45  components/draw/axes-1d.js
 13c19d4  35628 Nov 24 20:02  styles/app.scss
 fb0a662   2362 Nov 24 13:18  templates/components/app-header.hbs
 1b0d815    729 Nov 24 20:56  templates/components/draw/axis-menu.hbs