axis charts : allow results which don't cover the whole zoomedDomain
block-view.js : selectFeaturesCountsResults() : instead of filtering out results which do not cover the domain, prefer results which cover more of the domain.

models/block.js : add almostSubInterval() to select for display a featuresCounts result which covers enough of the zoomedDomain.
featuresCountsOverlappingInterval() : change from using inDomain (which returns a function not a value) to overlapInterval() or almostSubInterval()

data/block.js : getSummary() : calculate zoomedDomainText relative to size of zoomedDomain, so that it responds to change relative to the displayed interval.
zoomPanCalcs.js : rename overlapInterval{,1} and wrap it, to handle the case of interval completely contained within values.
interval-calcs.js : factor intervalOverlapCoverage() from (models/block.js) featureCountResultCoverage().
frontend/app/ :
 157b93b  7800 Dec  9 23:46  components/draw/block-view.js
 8be5afc 39722 Dec 10 11:17  models/block.js
 87102c6 48431 Dec  9 22:55  services/data/block.js
 02ed533 12162 Dec 10 11:17  utils/draw/zoomPanCalcs.js
 6b811f1  4049 Dec 10 11:20  utils/interval-calcs.js