graph width resizer : keep size stable on hover
 .. and drop out a couple of extra resizer icons so there is just 1.
app.scss : .resizer : to make the layout constant wrt :hover, adjust margin to compensate for the border 1px when hover, and don't vary font-size or font-weight.
Add margin-right: -7px to :hover (-6px - 1px to balance the border 1px); this aligns the icon above the border it moves, may drop this and the corresponding -6px) in a later commit : it looks less crowded without, and makes sense to be positioned left of the vertical border because it resizes the left element.
Move the padding : 2px so it is also constant - seems useful space.
Add to selector : not(.no-resizer-icon), so that the mapview.hbs resizer which brings its own glyphicon-resize-horizontal doesn't get an extra icon.

draw-map.hbs : comment out div.resizer;  this is adjacent a resizer in mapview which works, so this is probably not needed.
mapview.hbs : add .no-resizer-icon because this has a glyphicon-resize-horizontal (which looks better).
frontend/app/ :
 f89fa26 37052 Dec 18 20:52  styles/app.scss
 a3e95dc  1833 Dec 18 20:38  templates/components/draw-map.hbs
 e13e1ec  5515 Dec 18 19:51  templates/mapview.hbs