delay creation of popover until mouseover event
For large numbers of tracks, this was taking significant time, and the popover will only be used for a amall fraction of the tracks, so it is more efficient to create it when needed.
axis-tracks.js : factor showHorizTickHover out of configureHorizTickHover; change trigger (click hover) -> manual and add popover(show) and -hide.  use this.data() to check if already created.  Add mouseout : hideHorizTickHover to hide the popover.
 c72d01e 61875 Jan  6 19:26 frontend/app/components/axis-tracks.js