use primaryServer when session.requestServer not yet defined
This is for runtimeConfig which is used by getHoTLicenseKey() when the key is not defined in the build environment, so it gets it via API request from the runtime environment of the container.
runtimeConfig has no blockIds so blockServer is not defined, and buildURL() has not yet been called via adapters/application.js: updateRecord().
 c13176f 20632 Jan 20 12:41 frontend/app/services/auth.js