re-instate dependency on zoomedDomain change which was dropped in a2b67204
 ... this enables successive zooms to request more paths
models/block-adj.js
The change in a2b67204 models/block-adj.js to zoomedDomains() and domains(), to handle axis1d being undefined, dropped the facility of computed.mapBy() which has a dependency on the named field. So :
zoomedDomains() : change map to computed, so that dependency @each.zoomedDomainThrottled can be given.
domains : could make the equivalent change here, but instead aliased to axesDomains which is equivalent.
 f289dbc  23496 Apr 23 20:26 frontend/app/models/block-adj.js