add loadChr which splits a vcf and loads it in pieces to a single block
 ... for HighDensity SNPs - >1M per chr.
The first chunk is added using Datasets/createComplete which creates the dataset and block, and remaining chunks are appended to the block using Blocks/blockFeaturesAdd.
snps2Dataset.pl is used to convert the .vcf.gz to Pretzel json, which is piped to uploadData.

Copied from notes aws (02 - 5 Mar 2021) : datasetAndName2BlockId, dockerExec, DB_NAME

 54fcbfc2 3805 Apr 30 05:59 (UTC, in barley:tmp/) functions_data.bash (draft on local : 01988680 3655 2021-04-30 13:51)