axis tracks : use distinct yPosn and height for rects and triangles.
axis-tracks.js :
Add yPosnDir() which uses the y scale directly, and rename height() as heightDir() and add a wrapper height() which returns the absolute value.
Change the triangle functions to use those directional functions : featureTransform() : yPosn{,Dir} and triangleDimensions() : height{,Dir}.  The <rect>s are drawn with the non-directional equivalents yPosn() and height().  The change to height() in 1313065b : allow -ve result, to enable tracks triangles to have direction is not compatible with the requirements for <rect> tracks, so the 2 cases are now distinct.