re-connect rightEdge position update
axis-blocks.js : roll back to previous version (b220b866), before 5c215cb1 which factored component-layout out of axis-blocks so that it could implement centre & right in axis-blocks;  there is a problem with the dependencies through that layer of indirection, so instead of solving that atm, this commit rolls back this part of the change, because as yet we don't have any fixed width tracks in the right region.   This means that the functions moved to utils/draw/component-layout.js are now duplicates, and are not used.
axis-2d.hbs : to match that, append .centre in passing childWidths into axis-blocks