synteny blocks : update after zoom
draw-map.js :  axisApi : add reference to updateSyntenyBlocksPosition() and showSynteny(). add updateSyntenyBlocksPosition()
block-adj.js : updatePathsPosition() : if areSyntenyBlocks then updateSyntenyBlocksPosition().
Currently there is just a single result list : oa.syntenyBlocks, so this uses a task in draw-map which is singleton;  the results and the task can be moved into the corresponding block-adj-s.  When synteny blocks moves out to a separate component, the wiring via axisApi can be replaced.