style-editor : change bucketFill image for button and cursor
replace bucketFill.cur with a different image : bucketFill.png
app.scss : .style-editor > button : drop the width,height,padding-left which focused on the non-empty quadrant of the image.  add outline-style: dashed when hover, margin when not(.apply) to balance the .apply 1px border (so that toggle does not move the element).
README.md : add Acknowledgements, for the image bucketFill.png