synteny block size threshold : update the number display of the slider
draw-map.js and view-controls.js  : move sbSizeThreshold from oa. to controls.view. to allow update template dependency to update.
draw-map.js : updateSbSizeThresh() : drop -1 which does not apply to this .pow() calculation. Change approximate value to a full-precision calculation, to reach the endpoint 1e7, and use Math.E in place of 2.718.