graph height : reduce padding
app.scss : #holder : drop the padding added in d9b7c975, less available height when ! tablesPanelRight, and as needed this padding can be built into the svg viewBox so that the full space is used and not clipped.
axisTitleLayout.js : axisNameRows : reduce 7 -> 1, because the blocks list in the axis title will be moved into the axis title menu.
viewport.js : calc() : move topPanelHeight to apply only to documentElement.clientHeight, since it divHolder does not include it