Calculate the domain of the FC results when not merging.
After re-display of an uploaded block, lack of fcr.domain was causing features/chart to not display.
block.js : featuresCountsResultsMergeOrAppend() : when only 1 fcr or none to combine with (hence not using featuresCountsResultsMerge() which does featuresCountsResultsFilter()), use featuresCountsResultsDomain.
featuresCountsResults.js : add featuresCountsResultsDomain() : Calculate the domain of the results, from result extent.