explorer : active state of primary server tab
 ... was previously not cleared when switching the first time to another server tab
api-server.js : drop tabElement() (Lookup the tab <li> element in the data explorer corresponding to this API server), which was factored out of addClassActive(); not likely to use this other than possibly in some devel trace.
api-server-tab.js : factor firstTabActive() out of didRender(), use tab.select() in place of .addClassActive(), which is removed.
api-server-tab.hbs : factor (concat tab- apiServer.tabId) to this.tabId.