Improve on delay before collating ClientGroups.
This performs well on local;  timing on development server is different so this commit will be tested there.
client-groups.js : init() and whenModels() : trialled event listeners for : app, db, topology (t); no usable result - these can be dropped.
It was found that checking either ClientGroup.s.db or topology in poll is sufficient precursor to calling to ClientGroup.aggregate().