group edit page : add a back button;  select 1 of the 2 edit button icons : pencil
group.hbs : add link-to groups arrow-left;  indent the Group label from that slightly and increase its font-size.
drop 1 of the 2 buttons for editingName; keep the button-base and change its icon to pencil; the <button><i> works also and is similar complexity, but seems preferable to use the common component (standard / style guide).