vcfGenotypeLookup : increase height of textarea
move style textarea : width: 100% from axis-brush.hbs to app.scss, and add height: 20ex; overflow: auto.
These textarea-s are currently within the axis-brushes <table> column 1, which is a constraint on the width; not an issue because the vcf display is just a test / confirmation of the API and will be dropped as feature table / matrix-view are used.