Re-enable the VCF Lookup button
 ... which was affected by moving .brushedRegion to axis-1d in e1683595.
axis-1d.brushed was initially false and not updated, caused .brushedVCFBlocks to be [], and so brushedOrViewedVCFBlocks() returned .viewedVCFBlocks, which currently does not define .brushedDomain in the .axisBrush facade, and this caused vcfGenotypeLookupDomain to be undefined, so the VCF Lookup button was not displayed.

axis-1d.js : brushedRegion : initialising to null instead of undefined may assist setup of CP dependency (this change enabled the warning : .brushedRegion ... is being tracked by a tracking context ... mark the property as @tracked or use set(). )
axisBrush.js : brushHelper() : use .set() for axis1d.brushedRegion so that axis1d.brushed CP updates.