handle disparate axes length in filter pathIsSyntenic()
block-adj.js :
drawCurrentTask() : use (added) axesDomainSizes() instead of gradientThresholdToAxis(); the calculation is similar but instead of scaling pathGradient it passes the scale factor down to neighbourDistance() which applies it to each distance to produce a dimensionless distance.
Previously pathGradient was scaled up by the axis lengths of otherEnd, using gradientThresholdToAxis(); this didn't handle the addition of sortedEnd into the distance calculation.

add axesDomainSizes(), based on gradientThresholdToAxis(), which it can replace if this tests well.
add param domainSizes to pathIsSyntenic() and neighbourDistance().
neighbourDistance() : scale the path-neighbour value distance of each end of the path down by the axis length of that end.