genotype table : performance analysis
matrix-view.js :
createTable() : richSettings : afterScrollVertically() calls progressiveRowMergeInBatch(), so enable it with the same bit value as progressiveRowMergeInBatch().
Switch off #timeMeasure with trace variable - not needed in production.
updateSettings() : use bit value 0b100000 to enable progressiveRowMergeInBatch(), so only 1 value is used for it; in place of 0b1000000 which is now unused.

manage-genotype.js : haplotypeToggle() : use thenOrNow() for haplotypeFiltersSet() instead of local function filtersSet().
promises.js : thenOrNow() : handle undefined value, for use in haplotypeToggle() where textP may be undefined;  this is a change in functionality - a quick check of existing uses suggests they won't be affected;  possibly a separate function should be used for this case.