handle blast match type in gff
axis.js : axisConfig() : handle -ve endpoint of domain; this is an artifact of blast match in gff, and not valid for current data types (physical references and genetic maps), but it is reasonable to handle it if easy;  domainMax : apply abs to domain ends to get endMagnitudes, so that ratio > 1 and digits > 0, required for valid formatString.
gff_read.js :
featureParsed() : for type === match, extract scope from .attributes.description.
addFeature() : type === match : use .seq_id + .attributes.ID for name.