handle different name and scope in blast search
axis-ticks-selected.js : transientFeaturesLookup() : if block is not a key of featuresByAxis, find a key block which matches by scope or name and dataset.
Use serverSelected in place of primaryServer in blast-results-view.js : resultParentBlocks() and blast-results.js and mapview.js : viewDataset() and transient.js.
This could allow the results from one server to be viewed on a dataset of another server - which may or may not be a feature; alternately blast-results.js could record the dataset of the server on which the search is done, and pass its store to transient.