failed attempts to move column header height resizer into HandsOnTable
2 main approaches :
a. put the .resizer in the first column (Chr) : displays resizer icon (position is off) and receives drag; column headers are rotated so the dragging calculations are out of phase.
b. put it in the top-left corner .colHeader.cornerHeader; HandsOnTable overwrites this when re-render; icon not seen and no drag.

matrix-view.js :
add gtResizeHeaderHtml which is the html removed from matrix-view.hbs; using this directly didn't work so it was split into gtResizeHeaderClass and gtResizeHeader - similar result.
afterRender() : try later( 1000) for dragResizeListen().
colHeaders() : for first column (colIndex === 0) : colHeader : add gtResizeHeaderClass; try inserting gtResizeHeader also.
dragResizeListen() : resizeSel  : prefix with tableContainerSelector because it is now in table; narrow to .ht_master to avoid .ht_clone_top; b : try .colHeader.cornerHeader instead of .gtResizeHeader > div. Don't apply .filter() if dragResize is undefined.

matrix-view.hbs : remove .gtResizeHeader which is outside the table.
domElements.js : eltWidthResizable() : if resizer is not present (e.g. in case b), insert it. Works, and could be useful in other cases.