Apply ember-no-implicit-this-codemod
Apply ember-no-implicit-this-codemod. This joins a multi-line element onto a single line if it is not prettier-folded (1 arg per line), so apply prettier to these to re-fold them.  It also converts single- to double-quote, so revert those changes.

Use yapplabs version of package ember-radio-button.
Remove package ember-raf-scheduler, which is not updated for Ember v4.
utils/draw/axisBrush.js : comment out use of ember-raf-scheduler. axisScaleChangedRaf() : call fn directly instead of via scheduler.