handle reuse of feature name within a chromosome of a genotype dataset
matrix-view.js : featureValuesWidths : Chr : 35 -> 50 : increased to allow space for Chr prefix on chromosome name.
manage-genotype.js : brushedOrViewedVCFBlocksVisible() : add .filterBy(brushedDomain ) : only display and request genotypes on blocks which are on a brushed axis.
vcf-feature.js : addFeaturesJson() : append .value[0] to handle datasets with duplicate .name in 1 chr : added the position to the feature id, so that we can have the same name twice in one chromosome