ensure that Genotype Table is displayed when genotype-search is used
This is part of #385.
Use navigateGenotypeTable() to ensure that manage-genotype is displayed, hooked in after any input in genotype-search.

genotype-search.hbs : call action navigateGenotypeTable in a pipe with mut selectedDatasetId.

genotype-search.js :
 factor from sampleNameListInput{Key,}() to form setSamplesSelected(), and add setSamplesSelectedLater(), which is used instead.  This ensures that .manageGenotype is defined and displayed so that setSamplesSelected() can be done.
 add navigateGenotypeTable(), and use it in sampleNameListInputKey() and featureNameListInputKey() and vcfGenotypeSearch(), i.e. before the user gets to the Search button, .manageGenotype is defined and displayed; note that vcfGenotypeSearchDisabled does not update when .manageGenotype is defined.

mapview.js :
 setTab() : don't set if value is unchanged.
 window : add navigation : { }

 init() :
add .setTab to  .controls.window.navigation.
move this._super.apply(this, arguments)  from the end to the start of the function : the convention is definitely to call _super() at the start of init() and the other initialising life-cycle functions, and call at the end of the finalising functions such as willDestroyElement(); it seems to work at the end of init(), but this commit moves it for good practice.

table-dom.js : reduce trace