resultBlocks : when blockGiven return undefined instead of an empty Map.
manage-genotype.js : vcfGenotypeReceiveResult() : .resultBlocks is now undefined if ! blockGiven, so don't check .size.
vcf-feature.js : addFeaturesJson() resultBlocks : when blockGiven return undefined instead of an empty Map.
This description was given for genotype-search.js in 32bda852, which should have been : vcfGenotypeSearch() : use namesTrimUniq() in place of namesTrim() for .selectedFeaturesText and .selectedSamplesText