closes #398 : display blast errors in GUI to distinguish errors from empty results
feature.js : dnaSequenceSearch() : searchDataOut() : use ErrorStatus() to include status 400 in the error passed back.
blastServer.py : allow path of pretzel worktree to be configured - use os.getenv('pretzelDir') in place of /home/ec2-user/pretzel.
blastn_cont.bash : discard stdout of which blastn as only the status is required.

blastn_request.bash :
Handle ~/log being a soft link by appending /. in -d expression.
Add variable errorFile : blastn.error.
Move definitions of resultDir and resultFile before the first curl so they can be used there also. resultFile : use basename because  may be a path.
Introduce hostIpDefault to enable FLASK_PORT to be used to override the default 4000.
From feature/blastn_contPy 34652180 in .A3 : split up the curl command pipelines to access the exit status of curl.
If the curl result has a non-empty .error field, output it prefixed by 'Error:'.
Use --no-progress-meter instead of -s to suppress the progress-meter, so that error messages are output.

dnaSequenceSearch.bash : in defining blastn in the container case : if resourcesDir is not an absolute path, prefix blastn with serverDir; this was used in development, with servers running from the command line.