adjust mousewheel zoom at top of axis
manage-dataset.js : datasetOwned() : handle dataset not being populated with data fields.
zoomPanCalcs.js : wheelNewDomain() : mousePosition : offset mousePosition by a portion of the yOffset of the axis relative to the <svg>; this works ok on Chrome, not Firefox. The whole of wheelNewDomain() can be updated to use recent d3 features - the calculations date back to the original prototype.