genotype-search show HTML error text, ensure vcfGenotypeSamplesSelectedAll is defined, handle undefined lookupBlock.
genotype-search.js : navigateGenotypeTable() : get .manageGenotype.vcfGenotypeSamplesSelectedAll, which was seen to be undefined (i.e. .userSettings.vcfGenotypeSamplesSelected) - possibly manage-genotype was not yet created so userSettingsDefaults() had not been called.
manage-genotype.js :
showError() : use responseTextParseHtml() when error.responseText is defined instead of .responseJSON, as in components/form/base.js : handleError(), see comment there; see also checkError().
in vcfGenotypeLookup() and combineHeader() : use ? for .lookupBlock in the 2 remaining uses which did not have it (both are .hasTag); saw : this.lookupBlock is undefined, in combineHeader()